WIP: wifi: add support for Unisoc integrated SoC WiFi
Add a driver for Unisoc WiFi hardware with initial support for
integrated SoC WiFi using the SIPC-based host interface, tested on an
UMS9230 device with an UMW2631 wireless chip (an SC2355 derivative). The
driver is currently very basic and is missing features such as AP mode.

It should be possible to extend this driver to support other SC23xx-series
WiFi hardware from Unisoc with PCIe and SDIO host interfaces.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>