ASoC: sprd: add VBC V4 driver for DSP audio
Add a driver for normal audio playback/capture and voice call audio
using the DSP in Unisoc SoCs with VBC V4 hardware, such as UMS9230.
The actual VBC audio routing hardware is managed by the DSP, so the
driver only sends commands to the DSP and does not access the hardware
registers directly. For now, only implement a basic subset of the
routing and mixing controls supported by the DSP.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>