memory: mtk-smi: Add support for MT8163 SMI
Add support for SMI common and larbs on the MT8163 SoC.

The M4U on this SoC has 4 LARBs and one common SMI instance
without any sub common and gals.

The smi-common configuration is specific to this SoC. But, on the
LARB side, this is similar to that of MT8173, in the sense that it
doesn't need the port in LARB, and the register layout is also
compatible with that one, which allows us to re-use the smi-larb
platform data struct from MT8173.

Signed-off-by: bengris32 <bengris32@protonmail.ch>