mfd: mt6397: Use chip-specific interrupt count for domain creation
The IRQ domain is currently created with a hardcoded number of interrupts (MT6397_IRQ_NR),
which only matches the number of interrupts for the MT6397 PMIC.

However, this driver also supports other PMICs like MT6323, MT6328, and MT6331,
which have different (and in some cases, significantly higher interrupt counts).

As a result, interrupts beyond MT6397's limit are not usable on those chips.

Instead, pass the actual number of interrupts supported by the PMIC chip to
irq_domain_create_linear() based on the chip ID.

Signed-off-by: bengris32 <bengris32@protonmail.ch>