Input: goodix: Avoid overriding DT provided touch properties
The goodix driver allows the max_x and max_y properties to be
overriden from the device tree so long as the read config from the
touch controller is valid.

If that is not the case however, then the driver will override
max_x and max_y to default values regardless of whether they
are actually valid or not.

Fix this by checking again if max_x and max_y are invalid, and
if so then set them to safe defaults, but otherwise, use the
values parsed from the device tree.

Signed-off-by: bengris32 <bengris32@protonmail.ch>