arm64: dts: mediatek: mt8163-amazon-biscuit: Disable unused hardware
Disable the imgsys, mfgcfg, vdecsys and vencsys clock controllers on
biscuit as they are unused. Also disable the Mali GPU as it seems
to either be fused off or not present on the MT8163 variant used
on biscuit. Thermal is also disabled as the ADC sensors on this
variant aren't present either.

Signed-off-by: bengris32 <bengris32@protonmail.ch>