arm64: dts: mediatek: mt8163-amazon-biscuit: Disable unused hardware
Disable all display related hardware blocks on biscuit as they are
unused. Also disable the Mali GPU as it seems to either be fused off
or not present on the MT8163 variant used on biscuit.

This is done in a new mt8163-headless.dtsi file so other headless
MT8163 devices (should they exist) can include it in their device trees.

Thermal is also disabled on biscuit specifically as the ADC
sensors on aren't present.

Signed-off-by: Ben Grisdale <bengris32@protonmail.ch>