iio: light: tsl2584tsv: Fix resource leak and invalid free in probe
The usage of fail1 and fail2 labels are swapped. If iio_device_alloc
fails, then it jumps to fail1, which will actually call free() on
an invalid pointer. If either I2C communication or registering the IIO
device fails, then it jumps to fail2 which immediately returns without
freeing the iio_device.

Signed-off-by: bengris32 <bengris32@protonmail.ch>