mediatek: connectivity: Do not use stack-allocated buffers for DMA
Stack-allocated buffers are not guaranteed to be physically contiguous and cannot be
used with DMA.

The driver maintains a coalescing buffer that is allocated with kmalloc(GFP_KERNEL),
which provides physically contiguous memory suitable for DMA operations.

This prevents DMA mapping failures such as "rejecting DMA map of vmalloc memory" which happens
on kernels >= 5.4.

Change has been ported from the source code of the newer driver version [1].

[1]: https://github.com/amazon-oss/android_kernel_amazon_mt66xx-wifi/blob/lineage-18.1/module/core/gen2/common/wlan_lib.c

Signed-off-by: bengris32 <bengris32@protonmail.ch>